+++ cli/run.ts
@@ -1,21 +1,19 @@
 import { createSite } from "./utils.ts";
 
 interface Options {
-  root: string;
   config?: string;
 }
 
-export default function ({ root, config }: Options, scripts: string[]) {
-  return run(root, config, scripts);
+export default function ({ config }: Options, scripts: string[]) {
+  return run(config, scripts);
 }
 
 /** Run one or more custom scripts */
 export async function run(
-  root: string,
   config: string | undefined,
   scripts: string[],
 ) {
-  const site = await createSite(root, config);
+  const site = await createSite(config);
   console.log();
 
   for (const script of scripts) {